feat: add update method
